# openapi.yaml
components:
...
  parameters:
    BotName:
      description: The bot's name.
      in: path
      name: name
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BadRequestError"
      description: The request fails validation.
...
    NotFound:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/NotFoundError"
      description: The target does not exist or is not the caller's to act on.
...
    Unauthorized:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UnauthorizedError"
      description: No session cookie, or the session is expired or unknown.
  schemas:
    Accepts:
      description: The clocks the bot agrees to play.
      properties:
        match:
          type: boolean
        turnMs:
          description: The accepted turn time window [min, max] in milliseconds, or null for none.
          items:
            type: integer
          maxItems: 2
          minItems: 2
          nullable: true
          type: array
        unlimited:
          type: boolean
      required:
        - turnMs
        - match
        - unlimited
      type: object
...
    BadRequestError:
      properties:
        code:
          enum:
            - bad_request
          type: string
        error:
          type: string
      required:
        - error
        - code
      type: object
...
    BotLimitError:
      properties:
        code:
          enum:
            - bot_limit
          type: string
        error:
          type: string
      required:
        - error
        - code
      type: object
    BotListing:
      description: The declaration fields are absent until the bot declares them.
      properties:
        about:
          maxLength: 280
          type: string
        accepts:
          $ref: "#/components/schemas/Accepts"
        name:
          type: string
        online:
          description: True while the bot holds its stream open.
          type: boolean
        openForChallenges:
          description: True while the bot holds its stream open with open=1.
          type: boolean
        ownerName:
          nullable: true
          type: string
        provisional:
          $ref: "#/components/schemas/Provisional"
        rating:
          $ref: "#/components/schemas/Rating"
        repoUrl:
          anyOf:
            - format: uri
              maxLength: 2048
              type: string
            - enum:
                - ""
              type: string
        version:
          maxLength: 64
          type: string
      required:
        - name
        - ownerName
        - online
        - openForChallenges
        - rating
        - provisional
      type: object
    BotNameError:
      properties:
        code:
          enum:
            - invalid_name
            - name_reserved
          type: string
        error:
          type: string
      required:
        - error
        - code
      type: object
    BotToken:
      properties:
        name:
          type: string
        token:
          pattern: ^hxo_[A-Za-z0-9_-]{43}$
          type: string
      required:
        - name
        - token
      type: object
...
    InGameError:
      properties:
        code:
          enum:
            - in_game
          type: string
        error:
          type: string
      required:
        - error
        - code
      type: object
...
    NameTakenError:
      properties:
        code:
          enum:
            - name_taken
          type: string
        error:
          type: string
      required:
        - error
        - code
      type: object
    NotFoundError:
      properties:
        code:
          enum:
            - not_found
          type: string
        error:
          type: string
      required:
        - error
        - code
      type: object
...
    Provisional:
      description: True while the rating deviation is above 75.
      type: boolean
    Rating:
      description: The Glicko-2 rating, rounded to a whole point.
      type: integer
...
    UnauthorizedError:
      properties:
        code:
          enum:
            - unauthorized
          type: string
        error:
          type: string
      required:
        - error
        - code
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: opaque
      description: A bot token (hxo_...), minted and rotated by the owner on the website.
      scheme: bearer
      type: http
    gameToken:
      bearerFormat: opaque
      description: A game token (hgs_...) from a gameStart line.
      scheme: bearer
      type: http
    sessionCookie:
      description: An HttpOnly session cookie, set by the Discord OAuth callback or the guest route.
      in: cookie
      name: hexo_arena_session
      type: apiKey
info:
  title: HeXO Arena
  version: 0.19.0
openapi: 3.0.3
paths:
...
  /api/bots:
    get:
      description: The listed roster, ordered by name fold, without pagination. Delisted bots and bots of
        banned owners are hidden.
      operationId: listBots
      parameters:
        - description: Present as 1, narrows the roster to bots holding a stream open.
          in: query
          name: online
          required: false
          schema:
            enum:
              - "1"
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/BotListing"
                type: array
          description: The roster.
        "400":
          $ref: "#/components/responses/BadRequest"
      security: []
      summary: List public bots.
      tags:
        - Directory
    post:
      description: Mints the bot's token; this response is its only appearance. Names share one global
        namespace with users and never change.
      operationId: createBot
      requestBody:
        content:
          application/json:
            schema:
              properties:
                name:
                  pattern: ^[A-Za-z][A-Za-z0-9_-]{0,28}[A-Za-z0-9]$
                  type: string
              required:
                - name
              type: object
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotToken"
          description: The bot and its token.
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotNameError"
          description: The name fails the name syntax or is reserved.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotLimitError"
          description: The owner already holds 3 bots.
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NameTakenError"
          description: The name fold is taken by a user or a bot.
      security:
        - sessionCookie: []
      summary: Create a bot owned by the session user.
      tags:
        - Bots
  /api/bots/{name}:
    delete:
      description: "A bot with rated games is anonymized: it becomes a deleted-<n> placeholder, its games
        and ratings stay, and its name stays reserved. A bot without rated games is deleted and its
        name freed."
      operationId: deleteBot
      parameters:
        - $ref: "#/components/parameters/BotName"
      responses:
        "204":
          description: The bot and its token are gone.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InGameError"
          description: The bot is seated in a live game.
      security:
        - sessionCookie: []
      summary: Delete a bot owned by the session user.
      tags:
        - Bots
  /api/bots/{name}/token:
    post:
      description: The previous token stops working at once.
      operationId: rotateBotToken
      parameters:
        - $ref: "#/components/parameters/BotName"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotToken"
          description: The bot and its new token, shown only here.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
      security:
        - sessionCookie: []
      summary: Rotate a bot token.
      tags:
        - Bots
...